Return user paths relative to effective user's homedir
Since `os.path.expanduser()` is used to expand the `~` to the home
directory, paths which use `~` will always return a path relative to the
home directory of the user under which the python process is running.
However, if the euid has been changed, this means that these paths will
be incorrect, leading to problems when the non-privleged user tries to
write to these paths.

This commit fixes that issue by attempting to get the username matching
the euid, and using that username (e.g. `~user/.cache`) when invoking
`os.path.expanduser()`.