XDG_DATA_DIRS useage corrected, fixes wrong site_data_dir under Kubuntu
The freedesktop specification states that $XDG_DATA_DIRS is a preference ordered list of where data should be searched (as opposed to where applications should store data). Therefore, for multipath=False, which is used by applications to determine a path to store data, this can lead to unintended paths, as can be seen under issue #121. This commit changes the default directory in this case to the distribution independent '/usr/local/share'.